requirements.txt
slack-bolt
python-dotenv
openai
langchain
faiss-cpu
gradio
pandas
PyMuPDF
pillow